+++ components/org.wso2.carbon.identity.oauth.endpoint/src/main/java/org/wso2/carbon/identity/oauth/endpoint/util/factory/OAuth2AuthzServiceFactory.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.oauth.endpoint.authz;
+package org.wso2.carbon.identity.oauth.endpoint.util.factory;
 
 import org.wso2.carbon.context.PrivilegedCarbonContext;
 import org.wso2.carbon.identity.openidconnect.OpenIDConnectClaimFilter;
